+++ docs/test/maps.js
@@ -57,6 +57,11 @@ maps = {
     type: 'test',
     url: 'maps/test_single_base_features.json',
   },
+  zoom_details: {
+    name: 'Zoom Detail Features [360 bp]',
+    type: 'test',
+    url: 'maps/test_zoom_details.json',
+  },
   paper: {
     name: 'Paper Figure (E. coli PA2)',
     type: 'test',