+++ src/Slot.js
@@ -299,6 +299,7 @@ class Slot extends CGObject {
       const start = range.start;
       const stop = range.stop;
       if (this.hasFeatures) {
+        const drawnFeatures = [];
         let featureCount = this._features.length;
         if (!range.isMapLength()) {
           featureCount = this._featureNCList.count(start, stop);
@@ -324,13 +325,17 @@ class Slot extends CGObject {
           const sortedFeatures = this._featureNCList.find(start, stop, step).sort((a, b) => (a.score - b.score) || (a.length - b.length) );
           for (const feature of sortedFeatures) {
             feature.draw('map', slotCenterOffset, slotThickness, range, {showShading: showShading});
+            drawnFeatures.push(feature);
           }
         } else {
           // Draw by position (more efficient)
           this._featureNCList.run(start, stop, step, (feature) => {
             feature.draw('map', slotCenterOffset, slotThickness, range, {showShading: showShading});
+            drawnFeatures.push(feature);
           });
         }
+
+        this.viewer.annotation.drawFeatureLabels(drawnFeatures, slotCenterOffset, slotThickness, range);
 
         // Debug
         if (this.viewer.debug && this.viewer.debug.data.n) {
@@ -379,5 +384,4 @@ class Slot extends CGObject {
 }
 
 export default Slot;
-